+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2026-04-01
+
+### Added
+- `#wait_time(key)` for checking seconds until next allowed request
+- `SlidingWindow#window_reset_at(key)` for getting window expiry time
+
 ## [0.3.7] - 2026-03-31
 
 ### Added